Minor fixes for site rename revert Fixes to docs for plural s pipe, renamed file to match pipe
